add delete update işlemi satır üzerinde butonlar ile saglandı.

+++ server.py
@@ -302,69 +302,36 @@ def add_pic():
             pics.append(pic)
 
         connection.commit()
-
-
     if request.method =='POST':
-        picDesc = request.form['D']
+        if 'ADD' in request.form:
+            picDesc = request.form['ADD']
+            with dbapi2.connect(app.config['dsn']) as connection:
+                cursor = connection.cursor()
+                query = """INSERT INTO picPost (Description ) VALUES ('%s' )"""%picDesc
+                cursor.execute(query)
+                connection.commit()
+            return redirect(url_for('add_pic'))
 
-        with dbapi2.connect(app.config['dsn']) as connection:
-            cursor = connection.cursor()
-            query = """INSERT INTO picPost (Description ) VALUES ('%s' )"""%picDesc
-            cursor.execute(query)
-            connection.commit()
-        return redirect(url_for('add_pic'))
-    return render_template('add_pic.html', pics=pics)
+        if 'Delete' in request.form:
+            picId=request.form['id']
+            with dbapi2.connect(app.config['dsn']) as connection:
+                cursor = connection.cursor()
+                query = """DELETE FROM picPost WHERE PicId = '""" +picId + """'"""
+                cursor.execute(query)
+                connection.commit()
+            return redirect(url_for('add_pic'))
 
-@app.route('/delete_pic', methods = ['GET','POST'])
-def delete_pic():
-    pics = []
-    with dbapi2.connect(app.config['dsn']) as connection:
-        cursor = connection.cursor()
-        query = """SELECT picPost.PicId, picPost.Description FROM picPost"""
+        if 'Update' in request.form:
+            picId=request.form['id']
+            Desc = request.form['newname']
+            with dbapi2.connect(app.config['dsn']) as connection:
+                cursor = connection.cursor()
+                query = """UPDATE picPost SET (Description ) = ('"""+Desc+"""') WHERE PicId = '""" +picId + """'"""
+                cursor.execute(query)
 
-        cursor.execute(query)
+                connection.commit()
+            return redirect(url_for('add_pic'))
 
-        for pic in cursor:
-            pics.append(pic)
-
-        connection.commit()
-
-    if request.method =='POST':
-        picId = request.form['ID']
-
-        with dbapi2.connect(app.config['dsn']) as connection:
-            cursor = connection.cursor()
-            query = """DELETE FROM picPost WHERE PicId = '""" +picId + """'"""
-            cursor.execute(query)
-            connection.commit()
-        return redirect(url_for('delete_pic'))
-    return render_template('add_pic.html', pics=pics)
-
-@app.route('/update_pic', methods = ['GET','POST'])
-def update_pic():
-    pics = []
-    with dbapi2.connect(app.config['dsn']) as connection:
-        cursor = connection.cursor()
-        query = """SELECT picPost.PicId, picPost.Description FROM picPost"""
-
-        cursor.execute(query)
-
-        for pic in cursor:
-            pics.append(pic)
-
-        connection.commit()
-
-    if request.method =='POST':
-        picId = request.form['ID_update']
-        Desc = request.form['Desc']
-
-        with dbapi2.connect(app.config['dsn']) as connection:
-            cursor = connection.cursor()
-            query = """UPDATE picPost SET (Description ) = ('"""+Desc+"""') WHERE PicId = '""" +picId + """'"""
-            cursor.execute(query)
-
-            connection.commit()
-        return redirect(url_for('update_pic'))
     return render_template('add_pic.html', pics=pics)
